feat: (기능변경) N번마다 이동 전략 클래스(EveryNTimesMovableStrategy)를 2번마다 이동 전략 클래스(EveryTwiceMovableStrategy)으로 기능 변경
