+++ src/main/java/strategy/EveryTwiceMovableStrategy.java
